fixing some GCC compiler warnings
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1214 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ lass/spat/aabp_tree.inl
@@ -246,7 +246,7 @@ bool AabpTree<O, OT, SH>::intersects(
 
 		if (node.isLeaf())
 		{
-			for (int i = node.first(); i != node.last(); ++i)
+			for (size_t i = node.first(); i != node.last(); ++i)
 			{
 				LASS_SPAT_OBJECT_TREES_DIAGNOSTICS_VISIT_OBJECT;
 				if (TObjectTraits::objectIntersects(objects_[i], ray, tMin, tMax, info))
